move auto install
