+++ extension/src/webview/contract.ts
@@ -50,7 +50,8 @@ export enum MessageFromWebviewType {
   MODIFY_EXPERIMENT_PARAMS_RESET_AND_RUN = 'modify-experiment-params-reset-and-run',
   SET_EXPERIMENTS_HEADER_HEIGHT = 'update-experiments-header-height',
   INITIALIZE_PROJECT = 'initialize-project',
-  INSTALL_DVC = 'install-dvc'
+  INSTALL_DVC = 'install-dvc',
+  SETUP_WORKSPACE = 'setup-workspace'
 }
 
 export type ColumnResizePayload = {
@@ -185,6 +186,7 @@ export type MessageFromWebview =
   | { type: MessageFromWebviewType.SET_EXPERIMENTS_HEADER_HEIGHT }
   | { type: MessageFromWebviewType.INITIALIZE_PROJECT }
   | { type: MessageFromWebviewType.INSTALL_DVC }
+  | { type: MessageFromWebviewType.SETUP_WORKSPACE }
 
 export type MessageToWebview<T extends WebviewData> = {
   type: MessageToWebviewType.SET_DATA
